requirements.txt
pandas
openai
crewai
python-dotenv
google-generativeai
chromadb
streamlit
newsapi-python
